single_tech_samples/synapseanalytics/sample2_parking_sensors_infrastructure/tests/integrationtests/requirements.txt
pip
wheel
watchdog
flake8
coverage
Sphinx
twine
pytest
pytest-runner
tox
autopep8
azure-mgmt-resource
azure-mgmt-datafactory
azure-storage-blob
pyodbc
azure-synapse-artifacts
azure-identity
azure-storage-file-datalake
pandas